fedsfm_parser/requirements.txt
pandas
requests
beautifulsoup4
logging
python-telegram-bot==13.15
standard-imghdr
urllib3<2
